pageController에 그룹 페이지이동 추가 + Side Nav에 그룹페이지 경로 추가
