+++ src/main/resources/templates/group-details.html
@@ -533,7 +533,7 @@ Header END -->
               >
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="/group">
+              <a class="nav-link" href="/groups">
                 <i class="bi bi-people nav-icon"></i>
                 <span class="nav-text">Groups </span></a
               >
