+++ src/main/resources/templates/groups.html
@@ -531,7 +531,7 @@ Header END -->
                         >
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="/group">
+                        <a class="nav-link" href="/groups">
                             <i class="bi bi-people nav-icon"></i>
                             <span class="nav-text">Groups </span></a
                         >
