+++ src/main/resources/templates/index.html
@@ -591,7 +591,7 @@ Header END -->
               >
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="groups.html">
+              <a class="nav-link" href="/groups">
                 <i class="bi bi-people nav-icon"></i>
                 <span class="nav-text">Groups </span></a
               >
